refactor: move public routes into (public) group
Introduces the `(public)` route group so admin pages can render their
own chrome without CSS-hiding the shared Header/Footer. Every existing
public URL (/home, /FAQPage, /viewBrigadesPage, ...) stays identical —
route groups don't appear in the path.

- move every page under src/app/*/ into src/app/(public)/*/
- add src/app/(public)/layout.js rendering Header + Footer for the group
- shrink src/app/layout.js to html/body + Auth0 Providers
- promote providers.js from src/app/home/ to src/app/ so it wraps admin routes too
- forward NEXT_PUBLIC_AUTH0_AUDIENCE to Auth0Provider so admin can obtain
  an access token accepted by the backend
- update src/app/components/header.js to point at the moved menu/help buttons
