+++ src/app/(public)/FAQPage/components/FAQ.js
