+++ src/app/(public)/artigosPage/page.js
