+++ src/app/(public)/historiaPage/objetivosIcons.js
