+++ src/app/(public)/home/components/campaignsCarousel.js
