+++ src/app/(public)/home/components/campaignsCarousel.module.css
