+++ src/app/(public)/home/components/contactButton.js
@@ -1,22 +1,22 @@
-import Icons from "../../constants/icons";
+import Icons from "../../../constants/icons";
 import Image from "next/image";
-import styles from "./viewBrigadesButton.module.css";
+import styles from "./contactButton.module.css";
 import Link from "next/link";
 
-export default function ViewBrigades() {
+export default function Contact() {
   return (
-    <Link href="/viewBrigadesPage">
+    <Link href="/contactPage">
         <div className={styles.button}>
             <div className={styles.icon}>
                 <Image
-                    src={Icons.localizacaobranco.value}
-                    alt={Icons.localizacaobranco.alt}
+                    src={Icons.mail.value}
+                    alt={Icons.mail.alt}
                     height={21}
                     width={21}
                     />
             </div>
             <div className={styles.content}>
-                Conheça Nossas Brigadas
+                Entrar em contato
             </div>
             <div className={styles.icon}>
             <Image
