+++ src/app/(public)/home/components/viewBrigadesButton.js
@@ -1,22 +1,22 @@
-import Icons from "../../constants/icons";
+import Icons from "../../../constants/icons";
 import Image from "next/image";
 import styles from "./viewBrigadesButton.module.css";
 import Link from "next/link";
 
-export default function AddNewBrigade() {
+export default function ViewBrigades() {
   return (
-    <Link href="/contactPage">
+    <Link href="/viewBrigadesPage">
         <div className={styles.button}>
             <div className={styles.icon}>
                 <Image
-                    src={Icons.maisbranco.value}
-                    alt={Icons.maisbranco.alt}
+                    src={Icons.localizacaobranco.value}
+                    alt={Icons.localizacaobranco.alt}
                     height={21}
                     width={21}
                     />
             </div>
             <div className={styles.content}>
-                Cadastrar Nova Brigada
+                Conheça Nossas Brigadas
             </div>
             <div className={styles.icon}>
             <Image
