+++ src/app/(public)/home/homepage.module.css
@@ -1,7 +1,7 @@
 .firstSection {
     padding: 1.5rem 1.5rem;
     background-color: #466f34;
-    background-image: url('../icons/Artboard 4 1.png');
+    background-image: url('../../icons/Artboard 4 1.png');
     background-size: cover;
     background-position: center;
     background-repeat: no-repeat;
