+++ src/app/(public)/protectPage/logoutButton.js
@@ -1,4 +1,4 @@
-import Button from "../components/button";
+import Button from "../../components/button";
 import React from "react";
 import { useAuth0 } from "@auth0/auth0-react";
