+++ src/app/(public)/viewCampaignsPage/components/articleCard.js
