+++ src/app/(public)/viewCampaignsPage/components/articleCard.module.css
