+++ src/app/components/header.js
@@ -1,6 +1,6 @@
 import styles from "./header.module.css";
-import MenuButton from "../home/components/menuButton";
-import HelpButton from "../home/components/helpButton";
+import MenuButton from "../(public)/home/components/menuButton";
+import HelpButton from "../(public)/home/components/helpButton";
 import Link from "next/link";
 import Image from "next/image";
 import Icons from "../constants/icons";
