+++ src/app/layout.js
@@ -1,8 +1,6 @@
 import "./globals.css";
 import { Montserrat } from "next/font/google";
-import { Providers } from "./home/providers";
-import Header from "./components/header";
-import Footer from "./components/footer";
+import { Providers } from "./providers";
 
 export const metadata = {
   title: "Conexão Brigada",
@@ -15,15 +13,15 @@ const montserrat = Montserrat({
   subsets: ["latin"],
 });
 
+// Root layout stays minimal: html/body + Auth0 provider. Public and admin
+// route groups render their own chrome (header/footer or sidebar) so they can
+// diverge freely without CSS hacks. Every page still gets Auth0 context.
 export default function RootLayout({ children }) {
   return (
     <html lang="pt-BR">
       <body className={montserrat.className}>
-        <Header />
         <Providers>{children}</Providers>
-        <Footer />
       </body>
     </html>
   );
 }
-
